Completed cache functionality
Added listing, adding and removing features to the
caching subparser and completed the functionality for
these items and the clean and rebuild options.
Also cleaned the code from the cache file a bit.

+++ lib/jim_cache.py
@@ -16,25 +16,45 @@ import json
 import os
 import uuid
 
+import jim_config as config
+
+def _cache_name():
+    return '/var/cache/jim/jim-dependencies.json'
+
+def _cache_default():
+    return {}
+
 def _cache_error(message):
     print("jim: caching error: " + message)
     exit(1)
 
-def _cache_read():
-    # If there is no dependencies file create one
-    if not os.path.exists('/var/cache/jim/jim-dependencies.json'):
-        os.access(os.path.dirname('/var/cache/jim/jim-dependencies.json'), os.W_OK) \
+# Ensures the cache file is present
+def _cache_ensure():
+    _file = _cache_name()
+    # If there is no configuration file create one
+    if not os.path.exists(_file):
+        os.access(os.path.dirname(_file), os.W_OK) \
         or _cache_error("cache is not writable, something went wrong")
-        # Dump a blank json object into the file
-        with open('/var/cache/jim/jim-dependencies.json', 'wb') as fp:
-            fp.write(json.dumps({}))
+        # Dump a default json object into the file
+        with open(_file, 'wb') as fp:
+            fp.write(json.dumps(_cache_default() ) )
+
+# Returns the cache as an associative array
+def _cache_read():
+    _cache_ensure()
     # Load the file into data and return it
-    with open('/var/cache/jim/jim-dependencies.json', 'rb') as fp:
+    with open(_cache_name(), 'rb') as fp:
         data = json.load(fp)
     return data
 
+# Writes the associative array out to the cache
+def _cache_write(data):
+    with open(_cache_name(), 'wb') as fp:
+        fp.write(json.dumps(data))
+
+# Stores a file in the cache
 def _cache_store(data,_file):
-    print "No cache found, retrieving file: " + _file
+    print "no cache found, retrieving file: " + _file
     # Make a uuid name for the new file
     uuid_name = '/var/cache/jim/'+uuid.uuid4().hex+'.jim_cache'
     # Attempt to retrieve file and place in cache
@@ -44,12 +64,14 @@ def _cache_store(data,_file):
     data[_file] = uuid_name
     return data
 
+# Restores a file in the cache, file must have been stored already
 def _cache_restore(remote_file,cached_file):
-    print "Forced rebuild, retrieving file: " + remote_file
+    print "forced rebuild, retrieving file: " + remote_file
     # Attempt to retrieve file and place in cache
     os.system('curl -# '+remote_file+' -o '+cached_file) > 0 \
     and _cache_error("failed to download file for caching")
 
+# Gets a file name from the cache
 def _cache_get(_file,force):
     # Read in cache data
     data = _cache_read()
@@ -58,22 +80,50 @@ def _cache_get(_file,force):
         data = _cache_store(data,_file)
     # If the file was found but told to force, restore the file
     elif force:
-        _cache_restore(_file)
+        _cache_restore(_file,data[_file])
     # Save new cache data
-    with open('/var/cache/jim/jim-dependencies.json', 'wb') as fp:
-        fp.write(json.dumps(data))
+    _cache_write(data)
     return data[_file]
 
+# Rebuilds the entire cache
 def _cache_rebuild():
     # Read in cache data
     data = _cache_read()
     # Restore every file in the cache
     for _file in data:
         _cache_restore(_file,data[_file])
+    print "cache was rebuilt"
 
+# Clears the entire cache
 def _cache_clear():
     folder = '/var/cache/jim/'
     for _file in os.listdir(folder):
-        file_path = os.path.join(folder,_file)
-        if os.path.isfile(file_path):
-            os.unlink(file_path)
+        if _file != config._global_name():
+            file_path = os.path.join(folder,_file)
+            if os.path.isfile(file_path):
+                os.unlink(file_path)
+    print "cache was cleared"
+
+# Adds a specific file to the cache
+def _cache_add(_file):
+    _cache_get(_file,True)
+
+# Removes a specific file from the cache
+def _cache_remove(_file):
+    # Read in cache data
+    data = _cache_read()
+    if not _file in data:
+        print "file not found in cache"
+        return
+    os.unlink(data[_file])
+    del data[_file]
+    # Save new cache data
+    _cache_write(data)
+    print "file removed from cache"
+
+# Lists all files in the cache
+def _cache_list():
+    # Read in cache data
+    data = _cache_read()
+    for _file in data:
+        print _file
